[TimePicker]: both skins moved to uui, and loveship design has been changed into the promo implementation, and 48 size is deprecated.
