+++ app/src/docs/_examples/timePicker/TimeFormat.example.tsx
@@ -5,7 +5,7 @@ export default function TimePickerTimeFormatExample() {
     const [value, onValueChange] = useState({ hours: null, minutes: null });
 
     return (
-        <FlexRow>
+        <FlexRow rawProps={ { style: { minWidth: '175px' } } }>
             <TimePicker value={ value } onValueChange={ onValueChange } format={ 24 } />
         </FlexRow>
     );
