+++ epam-promo/components/inputs/index.tsx
@@ -1,5 +1,4 @@
 export * from './MultiSwitch';
 export * from './Rating';
 export * from './TextInput';
-export * from './TimePicker';
 export * from './Slider';
