+++ uui/components/inputs/TimePickerBody.tsx
@@ -1,7 +1,7 @@
-import css from './TimePicker.module.scss';
 import { TimePickerBody as uuiTimePickerBody, TimePickerBodyProps } from '@epam/uui-components';
 import { withMods } from '@epam/uui-core';
-import { ReactComponent as ArrowIcon } from './../icons/folding-arrow-24.svg';
+import css from './TimePicker.module.scss';
+import { ReactComponent as ArrowIcon } from './../../icons/folding-arrow-18.svg';
 
 export const TimePickerBody = withMods<TimePickerBodyProps>(
     uuiTimePickerBody,
